requirements.txt
numpy
pandas
matplotlib
yfinance
streamlit
mysql-connector-python